+++ shuup/importer/admin_module/import_views.py
@@ -55,8 +55,8 @@ class ImportProcessView(TemplateView):
             if filename.endswith("csv"):
                 mode = "csv"
             if self.importer_cls.custom_file_transformer:
-                return self.importer_cls.transform_file(mode, filename, supplier=self.supplier)
-            return transform_file(mode, filename, supplier=self.supplier)
+                return self.importer_cls.transform_file(mode, filename)
+            return transform_file(mode, filename)
         except (Exception, RuntimeError) as e:
             messages.error(self.request, e)